third_party/WebKit/LayoutTests/fast/preloader/script.html

<body>
<script>
if (window.testRunner)
    testRunner.dumpAsText();

window.addEventListener('DOMContentLoaded', function() {
    if (window.internals)
        console.log(internals.isPreloaded("resources/script1.js") ? "PASS" : "FAIL");
}, false);
</script>
This test requires DumpRenderTree to see the log of what resources are loaded.
<script src=resources/non-existant.js></script>
<script>document.write("<plaintext>");</script>
<script src=resources/script1.js></script>
