third_party/WebKit/LayoutTests/fast/preloader/script-expected.txt

CONSOLE MESSAGE: line 8: PASS
This test requires DumpRenderTree to see the log of what resources are loaded.

<script src=resources/script1.js></script>
